[artf4990] : Console bootstrap should gracefully recover where it can
This commit adds try/catch blocks around all of the *Service.bootStrap() calls in the BootStrap script.

While the system may not be very functional if one or another service fails to initialize, this will allow the console to try to continue.

* grails-app/conf/BootStrap.groovy
added try/catch blocks around all the service inits

git-svn-id: https://ctf.open.collab.net/svn/repos/svnedge/trunk@1398 03e8f217-bfc6-4b7c-bcb7-0738c91e2c5f