[artf4985] This revision fixes the broken method calls that expects a method call instead of a closure to retrieve the value of i18n messages in the functional/controller tests.
The error message is as follows:

Caused by: java.lang.IllegalStateException: No thread-bound request found: Are you referring to request attributes outside of an actual web request, or processing a request outside of the originally receiving thread? If you are actually operating within a web request and still receive this message, your code is probably running outside of DispatcherServlet/DispatcherPortlet: In this case, use RequestContextListener or RequestContextFilter to expose the current request.
	at com.collabnet.svnedge.console.ui.StatisticsController.<init>(StatisticsController.groovy:71)

* StatisticsController.groovy
- Solves the problem of the closure call that has no thread-bound.

git-svn-id: https://ctf.open.collab.net/svn/repos/svnedge/trunk@1404 03e8f217-bfc6-4b7c-bcb7-0738c91e2c5f