[artf4960] : Support LDAP authentication in console
This commit enables LDAP authentication to the Console if LDAP auth is configured for Apache/Svn. 

The implementation relies on a new virtualhost in Apache which is protected in the same way as /svn and /viewvc. To authenticate an LDAP user, the credentials are attempted at the protected location and, on success, the authentication token is created and ROLE_USER is granted. A shadow DB user of the same ID is also created, which can be added to additional console admin Roles. The existing Apache configuration was moved to it's own virtualhost, so that SSL could be applied independently of the auth helper.

* grails-app/conf/Config.groovy
added new property, "defaultApacheAuthHelperPort" for the protected virtual host

* src/groovy/../CsvnAuthenticationProvider.groovy
new acegi AuthProvider implementation that coordinates DB and LDAP authentication

* grails-app/services/../LifecycleService.groovy
added support for initializing Server.authHelperPort

* test/integration/../LifecycleServiceIntegrationTests.groovy (testAuthHelper)
new test to verify functioning of auth helper virtualhost in Apache.

* grails-app/i18n/messages.properties
new key for warning that LDAP users are not editable in the console

* grails-app/conf/spring/resources.groovy
new spring bean defined, csvnAuthProvider

* grails-app/conf/SecurityConfig.groovy
changed auth providers to use new csvnAuthProvider instead of builtins

* grails-app/domain/../Server.groovy
new nullable field, authHelperPort

* grails-app/services/../ServerConfService.groovy
changes to apache conf files to listen on auth helper port; moved configuration to virtualhosts to enabled simultaneous SSL and non-SSL.

* grails-app/views/user/show.gsp
conditional availability of "edit" buttons (LDAP users are not editable)

* grails-app/services/../UserAccountService.groovy (isLdapUser)
new method to assess whether a given User derived from LDAP auth

* grails-app/controllers/../UserController.groovy (showSelf, show, edit)
now tests whether a user is editable (isLdapUser) and configures view accordingly



git-svn-id: https://ctf.open.collab.net/svn/repos/svnedge/trunk@1449 03e8f217-bfc6-4b7c-bcb7-0738c91e2c5f