[artf5023] This revision fixes the problem related to the error and success messages being shown at the same time when saving the repo authorization access rules.
Another scenario of the error messages being displayed was when the user
clicked clicks in the Access Rules menu item right after an error message is
displayed.

* RepoController.groovy

- Adding "flash.clear()" to the "editAuthorization" action to not show any 
flash message, if any exists. This is related to the fact that the UI only 
renders errors that come from the action "saveAuthorization".

- Removing the reference to the flash.error and flash.message when there are 
no errors and when there are errors, respectively.

git-svn-id: https://ctf.open.collab.net/svn/repos/svnedge/trunk@1464 03e8f217-bfc6-4b7c-bcb7-0738c91e2c5f