[artf4970] Followup to r1478
Previous commit was incorrect and broke the publishing.  We should include
both the 32 and 64-bit versions of libraries.  SIGAR is smart enough to load
the correct library based on the OS and JVM.  A user can run a 32-bit JVM on
a 64-bit OS, in which case the 32-bit lib has to be loaded.  So we should be
including both.  We do this on 32-bit Linux too just for simplicity.

* scripts/Dist.groovy
  Always include 32/64 bit SIGAR libs

git-svn-id: https://ctf.open.collab.net/svn/repos/svnedge/trunk@1481 03e8f217-bfc6-4b7c-bcb7-0738c91e2c5f