[artf5038] : Server settings such as Port get saved, when there are validation errors
It looks like the call to lifecycleService to validate a low port causes the hibernate session to flush (because of automatic transaction wrapping on service methods?)-- so if Server instance has any changes, they persist. I changed the order of operations so that Server does not yet have form values applied to it when the service is consulted. 

* grails-app/controllers/../ServerController.groovy
validates params.port instead of server.port for standard port ok. the form params are afterwards applied to the Server instance.


git-svn-id: https://ctf.open.collab.net/svn/repos/svnedge/trunk@1554 03e8f217-bfc6-4b7c-bcb7-0738c91e2c5f