Update package prototypes for Solaris.  There will still be more to do.
The entries designated for os = unix do NOT get included on Solaris.  So we
need to duplicate them with os = solaris.  I am not sure if there is a
"not Windows" we could use or some other value that combines them.

The Solaris sparc libraries for sigar are still missing.  I do not know if
we can add an arch = argument here.  If not, then the Dist script will need
to be modified to include both the sparc and x86 sigar libraries in all
Solaris packages so that we can include them both.

git-svn-id: https://ctf.open.collab.net/svn/repos/svnedge/trunk@1571 03e8f217-bfc6-4b7c-bcb7-0738c91e2c5f