[artf5087] : Incorrect warning about httpd.conf
At some point, the way the confdirpath was put togehter changed, so that an upgraded system can have a good path to the includes, but forward/backward slashes might mismatch slightly on windows and the warning appears on the server page.

Fix was to verify the includes while ignoring differences in path separator.

* grails-app/services/ServerConfService.groovy (createOrValidateHttpdConf, verifyFileContains)
added a new method to verify file contents that takes a switch for ignoring differences in path separators.

git-svn-id: https://ctf.open.collab.net/svn/repos/svnedge/trunk@1614 03e8f217-bfc6-4b7c-bcb7-0738c91e2c5f