[artf5106] : i18n problems when installing updates
message bundle placehoders not expanding in the update progress messages, caused by single-quotes around the placeholders. Fix was to convert '{0}' to ''{0}'' in these messages (plus a couple of others with the same problem in the repository group).

* grails-app/i18m/messages*.properties
updated all languages to escape single quotes around placeholders, which are otherwise not expanded

git-svn-id: https://ctf.open.collab.net/svn/repos/svnedge/trunk@1619 03e8f217-bfc6-4b7c-bcb7-0738c91e2c5f