[artf5111] : Cannot run two versions of SVN Edge on same system due to port conflict
This bug is caused when 2 svn instances are running on the same host--the LDAP auth helper "listen" directive was always being written and was in conflict because the second could not distinguish the apache bound on the port to be it's own. While exploring this, it became clear there may be cases where ldap auth should not be allowed for the console even when it is permitted for apache, since SVN authz rules may be used to control access and the console does not use this. A new toggle, "Console LDAP authentication" was added to control this.

If 2 svn edge instances need distinct LDAP configuration, the authhelper port can be manually adjusted via new admin field (server settings -> auth helper port) is the work around for the bug.

* grails-app/conf/UpgradeBootStrap.groovy (release1_3_1)
new upgrade method to configure Server.ldapEnabledConsole field on existing dbs

* grails-app/controllers/../ServerController.groovy (editAuthentication)
will now initialize the authHelperPort for editing on the auth screen, if it's not already setup

* grails-app/domain/../Server.groovy
added 'ldapEnabledConsole' property

* grails-app/i18n/messages.properties
new keys for Console LDAP Auth fields

* grails-app/services/../ServerConfService.groovy
now conditionally writes the "Listen ${authHelperPort}"; consults new ldapEnabledConsole field; changed ServerMode comparisons in anticipation of Replica mode.

* grails-app/services/../LifecycleService.groovy (bootstrapServer)
initializes Server.ldapEnabledConsole field for clean install & in-mem db

* grails-app/views/servers/editAuthentication.gsp
added authHelperPort and ldapEnabledConsole fields to the server settings authentication edit screen.

* src/groovy/../CsvnAuthenticationProvier.groovy
now consults Server.ldapEnabledConsole before attempting to use the authHelper URL





git-svn-id: https://ctf.open.collab.net/svn/repos/svnedge/trunk@1631 03e8f217-bfc6-4b7c-bcb7-0738c91e2c5f