[artf5164] : Windows Installer: Provide an upgrade method
Modify the installer so that each new build generates a new product code
GUID.  This will causes Windows Installer to allow an upgrade to happen. The
upgrade it provides is an ugly bruter force uinstall/reinstall process, but
it works.  Customized the installer a little to allow it, but not much else
to do.

Also modified installer to take a System restore point which some users
mentioned in forums, and tweaked the dialog UI to include our product name
in place of "Advanced Installer"

git-svn-id: https://ctf.open.collab.net/svn/repos/svnedge/trunk@1652 03e8f217-bfc6-4b7c-bcb7-0738c91e2c5f