[artf5165] : Problems with CTF conversion functional tests
Current issue with "duplicate object" error may be caused by the test system not correctly cleaning up after itself. When there are errors in a conversion, the revert function is meant to be called -- but a signature mismatch from the localization effort is causing this to fail.

This fixes the invocation of "doRevertFromCtfMode" on convert error.

* grails-app/services/../SetupTeamForgeService.groovy (convert)
fixed faulty "doRevert" call



git-svn-id: https://ctf.open.collab.net/svn/repos/svnedge/trunk@1658 03e8f217-bfc6-4b7c-bcb7-0738c91e2c5f