[artf5173] : Problem editing users due to i18n
The user GSPs did not provide "action" attributes for the actionSubmit tags, so the value  was used to direct to controller action -- only worked for english once the value  was i18n. Added action attributes for all such tags.

Also removed ":" from "Change password" link on the edit user page via messages.properties. This is not really a field label, but rather an action that reveals the fields for editing. 

* views/repo/edit.gsp 
this page is not used, but added an action attribute anyway

* views/user/edit.gsp
* views/user/show.gsp
added action attributes

* i18n/messages*.properties
removed colon from "user.page.edit.passwdchange"

git-svn-id: https://ctf.open.collab.net/svn/repos/svnedge/trunk@1664 03e8f217-bfc6-4b7c-bcb7-0738c91e2c5f