[artf5124] Adding the support to retrieve the list of replicable external system for during the setup of the replica for a given TeamForge server. Each item of the list contains the properties of the external system: id, title, isSvnEdge, scm URL and Scm viewer URL. The confirmation page shows the title, the SCM URL and if the given server is SvnEdge or TeamForge SCM server.
This revision adds an updated version of the CollabNet SDK that exposes
the method Scm.getReplicableScmExternalSystems(), which is proxied in
the CtfRemoteClientService. 

* SetupReplicaController.groovy
- Retrieving the list of integration servers and getting the selected
one during the setup step. This is to be shown in the confirmation
page.

* messages.properties
- Adding labels for the form to identify if the external system to be
replicated is an svnedge server or a ctf one.
- Adding lables for the type of server (confirmation page)

* CtfRemoteClientService.groovy
- Adding the method getReplicableScmExternalSystemList()
that retrieves the list of Scm external systems from a CTF server.
- The properties of each server is placed in a List of Maps. Each map
in the list contains the id, title, isSvnEdge, scmUrl and scmViewerUrl.

* SetupReplicaService.groovy
- Refactoring the method call getIntegrationServers() to return the
result from the proxy call to the CtfRemoteClientService.

* confirm.gsp
- Adding the properties if the chosen server is SvneDge or
TeamForge.

* replicaSetup.gsp
- Refactoring the list of Scm External systems to display the title,
the type of the server (SvnEDge or TeamForge SCM) and the server
SCM URL.

* sf_soap50_sdk.jar
- Updating the sdk with the replication updated client
containing the Scm.getReplicableScmExternalSystemList().


git-svn-id: https://ctf.open.collab.net/svn/repos/svnedge/trunk@1690 03e8f217-bfc6-4b7c-bcb7-0738c91e2c5f