[artf5317, artf5245] This revision includes the refactoring necessary to fix the Replica Commands Executor. It also includes the most updated version of the TEamForge SOAP SDK.
The second part of the revision is the inclusion of the new command
ReplicaApproveCommand, which is used to update the state of the replica server
once it has retrieved the Approval/Denial command from the the Master server. 

A few classes were renamed to include the term "ReplicaCommand" instead of
"ActionCommand".

The upload of the command execution results are now the next step.

* BootStrap.groovy
- Adding the setupReplicaService in the bootstrap as the integration server 
setup. This is due to the fact that because the command executor needs a
reference to it as a Bean retrieved from the context object, as well as for
future use. The exception was 
"org.springframework.beans.factory.NoSuchBeanDefinitionException: No bean
named 'setupReplicaService' is defined"

* FetchReplicaCOmmandsJob.groovy
- Renaming the job.
- Refactoring the job class to reference to the renamed service
"replicaCommandExecutorService".

* CtfRemoteClientService.groovy
- REfactoring the method "getReplicaQueuedCommands" to retrieve and transform
the queued commands into internal representation of commands. This includes
the support to commands with parameters. 
- TODO: the previous use of the API was to provide the repositoryName as a
regular parameter. This should be changed on the EJB layer to better clean the
transport payload (as the repositoryName is NULL when not defined).

* ReplicaCommandExecutorService.groovy
- Renaming the service from the "ActionCommand".
- Refactoring the method "retrieveAndExecuteActionCommands" to retrieve the
queued commands from using the CTF client API.
- Removing the step to transform the command from the WS as it is already
transformed into List<Map> in the ctf client for reusability.
- Refactoring the method "processCommandRequest" to build the name of the
class with the correct package name.

* SetupReplicaService.groovy
- Adding the bootstrap method to the service.
- Adding the method updateServerAfterApproval(scmMasterUrl) that updates the
state of the DB with the Master SCM URL and the approval state to Approved.

* replication.command.*
- Moving the command classes to the new package.
- Adding the new command class "ReplicaApproveCommand.groovy" to process the
approval queued command. 

git-svn-id: https://ctf.open.collab.net/svn/repos/svnedge/trunk@1718 03e8f217-bfc6-4b7c-bcb7-0738c91e2c5f