[artf5334] : GRE on trying to converting to replica mode
The service method is rendering messages for the UI using a request locale, so Locale param needed to be added to the signature. 

In fixing this, i discovered that the test CTF (cu149) was not up to date and lacked the getReplicableExternalSystems() method -- so I added logic to handle this exception (and a UI message that the CTF host does not support replication).

* grails-app/controllers/../SetupReplicaController.groovy (replicaSetup)
now catches RemoteMasterException from the CtfRemoteService if the host does not have the requisite endpoint and displays error message

* grail-app/services/../SetupReplicaService.groovy (getIntegrationServers)
now extracts userLocale from ctfConnection bean and passes to CtfRemoteClientService delegate

* grail-app/services/../CtfRemoteClientServiceService.groovy (getReplicableScmExternalSystemList)
added locale param to signature

* grails-app/i18n/messages.properties
added key "ctfRemoteClientService.host.noReplicaSupport.error" for missing endpoints




git-svn-id: https://ctf.open.collab.net/svn/repos/svnedge/trunk@1727 03e8f217-bfc6-4b7c-bcb7-0738c91e2c5f