[artf5451] This revision fixes the wrong error message displayed when trying to convert a subversion edge server to replica with a user whose first credentials , the ones created by the system admin upon user creation, have not been changed by the user him/herself on CTF.
The message to be displayed is related to the error: "The credentials provided
need to be changed at the TeamForge server "http://ctf.server.url". 

After changing the password on CTF, the user can successfully try to create a
replica server.

* SetupReplicaController.groovy
- Retrieving the message key from the exception added to the
CtfAuthenticationException for the special cases.

* messages.properties
- Adding the new key "ctfRemoteClientService.auth.needschange" with the message
above asking the user to change the password at ctf.

* SetupTeamForgeService.groovy
- Just adding the same key of the exception into the creation of the new
instance with the new key.

* CtfRemoteClientService.groovy
- Verifying if the fault message contains the password error related to the
need to change the password.
- Using the new exception constructor with the key.
- While deleting the replica server, the user might have had his/her account 
reset and the password might need changes as well.

* CtfAuthenticationException.groovy
- Adding the messageKey property to the exception to track which authentication
failure originated the exception.

git-svn-id: https://ctf.open.collab.net/svn/repos/svnedge/trunk@1810 03e8f217-bfc6-4b7c-bcb7-0738c91e2c5f