[artf5369] This revision fixes the failure in registration process that happens when the replica server does not restart due to the fact that the server can't bind to the configured port after the registration has started.
There are two places where this error occurs: before the replica is created on
the CTF manager, and after the registration process with CTF is finished and
the local approval command processes the command. In the former case, the 
registration process must stop and send the user back to the initial page,
while in the latter the approval command must pass instead of failing and
report the user in the status page.

* SetupReplicaController.groovy
- Verifying if there is any error in the initial of the process. This is the
error processed during the attempt to register the replica.
- Capturing the exception from the registration confirmation and "storing" it
at the ReplicaConversionBean.
- Showing the new error key for general errors.

* StatusController.groovy
- Show the error related to the server status if the user is using an approved
replica whose Apache server hasn't started successfully.

* messages.properties.groovy
- Adding the key "replica.error.registration.serverCantRestart" for the
registration failure during the approval processing.
- Adding the key "replica.error.registration" with the general error message.

* LifecycleService.groovy
- Adding the status if the server failed during the restart methods.
- hasFailedToRestart() gives the last status of restarting the server based on
the restart() methods.
- updating the startOrStopServer() with the methods "serverHasFailedToRestart()"
and "serverHasSuccessfullyRestarted()" to signal the server status.

* SetupTeamForgeService.groovy
- Adding the clause "throws CantBindPortException" to the method 
"restartServer()", to avoid the UndeclaredThrowableException when the method
fails with a CantBindPortException.

* SetupReplicaServer.groovy
- Adding the CantBindPortException to the method registerReplica throws list.
- Moving the call "setupTeamForgeService.restartServer()" from the
registration to the Approval Command.

* ReplicaConversionBean.groovy
- Adding the registrationError holder to show the error message during
the replica registration.

* ReplicaApproveCommand.groovy
- Try/Catch the CantBindPortException of the approval process to restart the
server and not fail.

git-svn-id: https://ctf.open.collab.net/svn/repos/svnedge/trunk@1814 03e8f217-bfc6-4b7c-bcb7-0738c91e2c5f