[artf5111] : Cannot run two versions of SVN Edge on same system due to port conflict
This commit addresses the issues reported in post12039 related to the auth helper port.

1) we now format the number without the comma
2) we now limit the value to a legitimate port number (1-65535)

This involved some changes to the GSP, the Server contstraints, and the message bundle.

* grails-app/domain/../Server.groovy (constraints)
added rules for authHelperPort (dependency on ldapEnabledConsole, max integer for port)

* grails-app/i18n/messages.properties
new messages related to authHelperPort validation

* grails-app/views/server/editAuthentication.gsp
changed the input field for ldapServerPort and authHelperPort so that the number is not formatted with comma

git-svn-id: https://ctf.open.collab.net/svn/repos/svnedge/trunk@1829 03e8f217-bfc6-4b7c-bcb7-0738c91e2c5f