[artf5496] : Problem when converting from replica mode to standalone
follow up commit: now archiving any existing local Repository during Replica Mode conversion. A warning message is displayed on the final page of the wizard.

* grails-app/services/../SetupReplicaService.groovy (registerReplica)
when all steps of the conversion are successful, the existing Repositories are removed

* grails-app/i18n/messages.properties
new warning message that local repos will be archived

* grails-app/controllers/../SetupReplicaController.groovy (confirm)
the warning message is flashed on this screen



git-svn-id: https://ctf.open.collab.net/svn/repos/svnedge/trunk@1830 03e8f217-bfc6-4b7c-bcb7-0738c91e2c5f