[artf5204] : Do not expose LDAP password in View Source
LDAP Bind Dn password was written clear in the html source for the password field. With this change, the password is obfuscated and the db value is updated only when the field is changed.

* grails-app/controllers/../ServerController.groovy (update)
now excludes LdapBindDNPassword from updating DB unless there is a change, so that obscured value is not saved

* grails-app/services/../SecurityService.groovy (generateAlphaNumericPassword)
new random password generator that is alpha-only (existing method used all ascii, so the "dots" in the password field didn't match length)

* grails-app/views/server/editAuthentication.gsp
switched password input to use new custom tag, g:passwordFieldWithChangeNotification

* grails-app/taglib/CsvnTagLib.groovy
new tag,  passwordFieldWithChangeNotification, that renders an obscured password field and a hidden field to flag user change

git-svn-id: https://ctf.open.collab.net/svn/repos/svnedge/trunk@1886 03e8f217-bfc6-4b7c-bcb7-0738c91e2c5f