[artf5451] Inappropriate alert message is displayed when a user without logging in to CTF tries to convert to replica
Change the message to more closely match that which is returned from TeamForge:

  "Please change your TeamForge password. Your password was set by an admin and
  it needs to be changed by logging into {0}."

Note that an early msg attempt tried to use the username in the message rather
than the URL, but this was stymied by behavior in SetupReplicaController.  As I
explain in a FIXME in that file, we may want to consider adding msg params to
the Exception classes, and use them instead of just assuming that URL params
will be used for all error messages.

grails-app/i18n/messages.properties

 * Update ctfRemoteClientService.auth.needschange with the new message.

grails-app/controllers/com/collabnet/svnedge/controller/integration/SetupReplicaController.groovy

 * Add the FIXME described above.


git-svn-id: https://ctf.open.collab.net/svn/repos/svnedge/trunk@1896 03e8f217-bfc6-4b7c-bcb7-0738c91e2c5f