[artf5724] This revision adds the runningCommandIds to the remote SOAP client method scm.getReplicaQueuedCommands().
As new commands are queued, a new associated instance of CommandResult is 
created with the commandId prior the command's scheduling/execution. In order
to support the new requirement, the list of CommandResults that haven't been
transmitted is added as the new parameter. This list is retrieved from the
CommandsResultDeliveryService.

* CommandResult.groovy
- Just fixing the Javadocs.

* FetchReplicaCommandsJob.groovy
- Retrieving the list of running/scheduled commands to be used as a parameter
to the refactored method signature of ctfRemoteClient.

* CtfRemoteClientService.groovy
- Adding the new parameter to the method getReplicaQueuedCommands() as the set
of instances of CommandResult.
- Building the Array of command Ids from the received list.

* CommandResultDeliveryService.groovy
- Adding the method "getUnacknowledgedExecutingCommandResults()" that returns
all the command results that haven't been acknowledged.

* sf_soap60_sdk.jar
- Adding the new version of the sdk with the new version of the method
scm.getReplicaQueuedCommands().

git-svn-id: https://ctf.open.collab.net/svn/repos/svnedge/trunk@1902 03e8f217-bfc6-4b7c-bcb7-0738c91e2c5f