[artf5584] This revision tries to increase the default thread pool size used by the backgroundService. The default value is set as 5 and that might be the cause of having command results being published at the same time, when threads that are processing repositories finish.
The replica command execution uses the global background thread service to
spawn threads to do the tasks of scheduling, executing and reporting results.
The default configuration of the backgroundService is to queue 1000 tasks on
a pool of only 5 threads. This revision attempts to increase this value to 60,
as a factor of 20 threads spawn by the 3 services defined in the command
execution.

* Config.groovy
- Changing the default threadCount for the background service from 5 (default)
to 60 as a factor of 20 threads for 3 services.

git-svn-id: https://ctf.open.collab.net/svn/repos/svnedge/trunk@1906 03e8f217-bfc6-4b7c-bcb7-0738c91e2c5f