[artf5743] This revision "beautifies" the laucher.html page that is displayed to windows users when the server is loading the server.
The two versions of the file launch.html for windows 32 and 64 bits were 
updated and they are the same. The launch.html will displays 5 different status
messages for about 30 seconds. The last status is displayed until:

a) the server is started, and the green "Ready" is displayed.
b) the server never starts, and the red message "check the logs" is displayed.

The messages are displayed using the viewvc CSS in addition to the same styles
used by the console as it was copied. The spinner icon is the same from the
software updates, but the new database images are suggested copies from public
sites.

* windows/installer|installer64/launch.html
- Adding the viewvc CSS as they are statically available.
- Updating the javascript to keep track of the load phase and total number of 
seconds the script tries to load the image URL.
- Adding the logo from viewvc code snippet header template
- Adding the footer from the viewvc code footer template.
- Adding status table template to display the message and the spinner.
- Adding the database image in the left-side of the page.

* styles_new.css
- Style needed for the inner status table. It was completely copied from the
application server.

* roller.gif
- The spinner copied from the packages update, with the background color
updated with the blue used by the header.

* database-down-icon-hi.png
* database-up-icon-hi.png
- The database icons for the server up and down statuses.

git-svn-id: https://ctf.open.collab.net/svn/repos/svnedge/trunk@1957 03e8f217-bfc6-4b7c-bcb7-0738c91e2c5f