[artf5777] : Cannot convert to replica mode after converting back from replica mode
When reverting, a ReplicaUnregisterCommand is created for undoing the replica configuraiton. It was successful, but the "replicaServer" category remained blocked, so that subsequent ReplicaApproveCommand (from the subsequent convert) could never run. Issue was simply the pseudo-ID of the UnregisterCommand was type Integer and needed to be type String.

* grails-app/services/.../CtfRemoteClientService.groovy (getReplicaQueuedCommands)
now prepends COMMAND_ID_PREFIX to the random id so that the mutex hashmap could work properly (requires String id)

git-svn-id: https://ctf.open.collab.net/svn/repos/svnedge/trunk@1961 03e8f217-bfc6-4b7c-bcb7-0738c91e2c5f