[artf5797] : svnedge: implement revprop sync command
Added a command implementation class and integration test for a new "CopyRevprops" command (the commandline invocation is "svnsync copy-revprops"). The test validates that an svn:log message updated from a master working copy on a particular revision will be copied to the replica working copy as expected.

* src/groovy/../CopyRevpropsCommand.groovy 
the command implementation 

* test/integration/../CopyRevpropsCommandIntegrationTests.groovy
the test case

git-svn-id: https://ctf.open.collab.net/svn/repos/svnedge/trunk@2006 03e8f217-bfc6-4b7c-bcb7-0738c91e2c5f