[artf5370] Sync fails if SVN server uses SSL
I was misreading the error messages on jenkins; the fault for StatusControllerTests was emanating from line StatusController l. 211. This commit suggested by Jmcnally should fix.

* grails-app/controllers/../StatusController.groovy (index)
logic to evaluate CTF cert (for second time in this method!) was not conditioned properly, since the "replicaConfiguration" instance appears to exist in some cases where the server is not configured as replica.


git-svn-id: https://ctf.open.collab.net/svn/repos/svnedge/trunk@2017 03e8f217-bfc6-4b7c-bcb7-0738c91e2c5f