[artf5552] This revision adds the initial implementation of the RealTimeCommandLineService, responsible to execute a command in the background and provide clients a Listener/handler to receive the command output lines.
* CommandLineService.groovy
- Organizing the imports.
- Adding the token COMMAND_TERMINATED to be captured during the asynchronous 
execution of commands. Value = "0xdeadc0de".
- Refactoring white spaces.
- Adding the new method "executeWithLineListener(command, queue), that will
execute the command in the background, while the process output in the standard
out will be offered to the given queue. This queue is used by the new class
RealTimeCommandLineService as a blocking queue consumer.

* RealTimeCommandLineService.groovy
- Adding the new service that can capture the output of a command. The output
is offered to the instance blocking queue and uses the above described token
COMMAND_TERMINATED.
- Adding the method 

git-svn-id: https://ctf.open.collab.net/svn/repos/svnedge/trunk@2044 03e8f217-bfc6-4b7c-bcb7-0738c91e2c5f