[artf5552]: Create Jenkins-like displays of the runtime output of the running replica commands in the web UI
This commit adds a "tail" action to the LogController and some frontend javascript for streaming in a logfile as it is written. The idea is that our ReplicaCommand executions will send their stdout to a log file which this can be used to read to a div. The command executions will need to delete their log files on completion.

I added a "tail" button the /log/show view which uses this functionality for the currently displaying log file -- we can remove if not needed, just was easier to develop the feature here and seems to have some value.

* grails-app/controllers/../LogController.groovy (tail)
new method to retrieve log file content from a given offset in JSON format

* grails-app/views/../show.gsp
added JS support and "Tail" button for streaming in the current  log

* grails-app/i18n/messages.properties
new keys for tail button

* web-app/js/application.js
new LogStreamer class for doing the ajax calls to LogController. Supports multiple intances per page pointing to different logs.


git-svn-id: https://ctf.open.collab.net/svn/repos/svnedge/trunk@2052 03e8f217-bfc6-4b7c-bcb7-0738c91e2c5f