[artf5552]: Create Jenkins-like displays of the runtime output of the running replica commands in the web UI
Upped the buffer size on the CommandLineService.execute method used for capturing output to file -- a ByteArrayOutputStream is used to "tee" the output for returning to the caller and writing file, and with no parameter, the byte array gets resized every 32 bytes. Now it is 512.

* grails-app/services/CommandLineService.groovy (execute(List command, Outputstream out, OutputStream err))
upped the ByteArrayOutputStream size from default (32) to 512.

git-svn-id: https://ctf.open.collab.net/svn/repos/svnedge/trunk@2098 03e8f217-bfc6-4b7c-bcb7-0738c91e2c5f