[artf5551] This revision finally fixes the list of jobs page for the replica status service.
Since the cometd messages are sent in a very small interval for 
short-running commands, there is no way to use timers on Javascript and 
speed down the UI changes in the browser. In addition, Javascript is 
single-threaded and, as a consequence, no way to use cometd callback method
with different method calls. For this reason, a handler was added to delay the
delivery of messages in intervals instead of submitting the messages as the
events about commands happen.

Other small changes include: 
- The size of the running tables, which reflects the
number of permits for short-running and long-running commands;
- Using one single cometd channel simplifies the UI handling;
- Adding a new event that is triggered just before the execution of a command.
- Removing the spinners from the running tables.

* JobController.groovy
- Adding the model values "maxLongRunning" and "maxShortRunning" as the values
configured for the replica server.

* messages.properties
- Adding descriptions for each command to be displayed in the screen.

* ReplicaServerStatusService.groovy
- Organizing imports for the changes.
- Using the background thread plugin to instantiate the new handler.
- Renaming a few properties.
- Adding a Blocking Queue to hold the delayed transfer of the events related to
the change of state of commands.
- Adding the helper static class CommandAtState, that represents the state of
a given command for a given event. The reason why is that the command's event 
property may change during the time events are triggered and the time they are
reached and published by this service. So, this form of index is safely used in
the transfer queue to publish the events through cometd.
- Simplifying the channels by consolidating the total number of commands
running into the messages.
- Removing the method publishCommandTransition() to be added to the new Handler
class.
- Fixing the method updateOrRemoveCommandState() that did not handle the 
indexes for long-running and short-running commands correctly.
- Replacing the event handling CommandReadyForExecutionEvent with the new
command "CommandAboutToRunEvent". This new event is triggered when a command
acquires a permit to run.

* ReplicaCommandExecutorService.groovy
- Triggering the new event CommandAboutToRunEvent before the command starts
running. This is needed because the list jobs page was showing the running
commands, even if they had been blocked because of permits. In this way, this
event will only be triggered when a command acquires a permit to run.

* _runningCommands.gsp
- Trying to use the new command description for commands whenever possible.
- Using the new method "getStateTransitionTime()".

* list.gsp
- Adding Javascript documentation
- Exposing non-cachable variables in the page instead of the javascript.
- Setting the maxNumber properties in both the long-running and short-running
commands tables.

* AbstractCommand.groovy
- Showing the state of the command on toString().
- Removing the running transition state from the run method in favor of using
it before the new event "CommandAboutToRunEvent" is triggered.
- Adding the methods getStateTransitionTime(), getStateTranstionTimes(), 
getStateAtTime().

* CommandAboutToRunEvent.groovy
- New event triggered when a command has acquired a permit and is about to run.
The previous event was triggered when a command was selected to be run, but 
the command still needed to acquire a permit, giving the wrong impression in
the status page when the max number of permits have been reached and commands
had been scheduled to run.

* CommandReadyForExecutionEvent.groovy
- Updating the Javadocs with the correct semantics of this event as described
above.

* CommandStateDelayedNotifierHandler.groovy
- Adding a new notifier handler for the status service that will only publish
a command state change event at every 4 seconds. It keeps a Linked Blocking
Queue for the sake of maintaining the order of the events.

* jobs-list.js
- Removing the channel not used and using the new one.
- Removing the update spinners methods.
- Refactoring the updateCommands() method.
- Fixing the method "highlightTerminatedCommand()" to correctly change the
color of the rows of terminated command in green/red depending on the command
result.

TODO:
- The UI can change some properties based on the command type. For instance,
the total number of permits can resize the running commands tables without 
reloading.

git-svn-id: https://ctf.open.collab.net/svn/repos/svnedge/trunk@2108 03e8f217-bfc6-4b7c-bcb7-0738c91e2c5f