[artf6034] : Command logging feature should not include passwords
Rolling back change to CommandLineService that recorded command tokens (including password) in the command output log. This was an attempt to validate the output stream before handing to the Process wrapper, but even streams that worked for writing the shell command tokens would subsequently fail with the stdout of the process. 

* grails-app/services/../CommandLineService.groovy (execute (List command...))
rolled back output stream validation logic which was writing shell command passwords to the file

git-svn-id: https://ctf.open.collab.net/svn/repos/svnedge/trunk@2163 03e8f217-bfc6-4b7c-bcb7-0738c91e2c5f