[artf5798] : develop tests to fully test command threading system
follow up commit to (hopefully) correct failures

* test/integration/../CommandThreadingIntegrationTests.groovy (testBasicBlockingAndConcurrency)
removed test comparing first and third repo, since it's not known the order in which commands are selected. now only tests commands for 2 repos.

* test/integration/../CommandThreadingIntegrationTests.groovy (testLogExecutionWithWrongParametersException)
fix for failure -- command logging will now fall back to ConfigurationHolder if the command context is null (tests where command is not instantiated by the framework), but this broke one test


git-svn-id: https://ctf.open.collab.net/svn/repos/svnedge/trunk@2217 03e8f217-bfc6-4b7c-bcb7-0738c91e2c5f